requirements.txt
vkbottle==4.3.10
discord~=2.0.0
aiogram~=2.22.1
aiohttp~=3.8.1
requests~=2.28.1
beautifulsoup4~=4.11.1
Pillow~=9.2.0
python-dotenv~=0.21.0
pytz~=2022.2.1